refactor(CastRow): decompose into primitives, add component test infra
Split src/common/components/CastRow.tsx (~1000 lines) into focused
primitives under CastRow/: Author, CastText, CastTime, EmbedSection,
ReactionBar, RecastBadge, linkify. StandardCastRow composes them.
CastRow.tsx is now a 2-line re-export aliasing StandardCastRow as
CastRow so existing callers work unchanged.

CastRowLegacy.tsx snapshots the pre-refactor file for the regression
test that compares old vs new render output. Slated for removal once
the new structure is established.

Adds @testing-library/react + jest-environment-jsdom for component
testing. Helper tests keep testEnvironment: node; component tests
opt into jsdom via per-file /* @jest-environment jsdom */ directive.
