+++ jest.config.js
@@ -2,7 +2,16 @@
 module.exports = {
   testEnvironment: 'node',
   transform: {
-    '^.+\\.tsx?$': ['ts-jest', {}],
+    '^.+\\.tsx?$': [
+      'ts-jest',
+      {
+        // Override tsconfig "jsx": "preserve" so JSX is compiled for Jest.
+        // Existing helper tests don't use JSX, so this is a no-op for them.
+        tsconfig: {
+          jsx: 'react-jsx',
+        },
+      },
+    ],
   },
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1',
@@ -13,5 +22,6 @@ module.exports = {
   testPathIgnorePatterns: [
     '/node_modules/',
     '/supabase/functions/', // Deno-based tests, run separately
+    '/__tests__/fixtures/', // shared fixture data, not test files
   ],
 };
