+++ package.json
@@ -191,6 +191,8 @@
     "@next/bundle-analyzer": "^16.1.5",
     "@next/eslint-plugin-next": "^13.5.8",
     "@tanstack/react-query-devtools": "^5.91.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
     "@types/jest": "^29.5.14",
     "@types/lodash.clonedeep": "^4.5.9",
     "@types/lodash.debounce": "^4.0.9",
@@ -216,6 +218,7 @@
     "eslint-config-next": "^14.2.35",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "kysely": "^0.27.6",
     "kysely-codegen": "^0.15.0",
     "lint-staged": "^16.2.7",
